Implemented basic NMS checking. Continuing on adding Island levels.
