+++ src/main/java/me/saber/skyblock/SaberSkyBlockAPI.java
@@ -1,6 +1,6 @@
 package me.saber.skyblock;
 
-import me.saber.skyblock.Island.IslandUtils;
+import me.saber.skyblock.island.IslandUtils;
 import me.saber.skyblock.filemanager.FileManager;
 import me.saber.skyblock.generators.WorldGenerator;
