+++ src/main/java/me/saber/skyblock/events/IslandEvents.java
@@ -1,6 +1,6 @@
 package me.saber.skyblock.events;
 
-import me.saber.skyblock.Island.Island;
+import me.saber.skyblock.island.Island;
 import me.saber.skyblock.Main;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
