+++ src/main/java/me/saber/skyblock/filemanager/FileManager.java
@@ -8,14 +8,14 @@ import java.io.File;
 public class FileManager {
 
     public CustomFile dataFileCustom = new CustomFile(new File(Main.getInstance().getDataFolder()+"/data.yml"));
+    public CustomFile levelWorth = new CustomFile(new File(Main.getInstance().getDataFolder()+"/level-worth.yml"));
 
-    public File dataFile = dataFileCustom.getFile();
     public FileConfiguration d = dataFileCustom.getFileConfig();
 
     public void setup(){
 
         dataFileCustom.setup(false);
+        levelWorth.setup(true);
 
     }
-
 }
